--- public/style.css
@@ -1,5 +1,0 @@
-html,
-body {
-  margin: 0;
-  padding: 0;
-}
