+++ src/index.tsx
@@ -1,6 +1,19 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
+import { createGlobalStyle } from 'styled-components';
 import App from './components/App';
 
+const GlobalStyle = createGlobalStyle`
+body {
+  margin: 0;
+  padding: 0;
+}
+`;
+
 const root = createRoot(document.getElementById('root')!);
-root.render(<App />);
+root.render(
+  <React.StrictMode>
+    <GlobalStyle />
+    <App />
+  </React.StrictMode>,
+);
